Return only the raw evaluation error message without prefix
At the moment we store any evaluation error as `string(err)` which is
great for its simplicity and allows us to unwrap the error and test it
with `errors.Is` or `errors.As`, but the string representation also
includes the string representation of the error itself.

In order to both support unwrapping but also be able to extract only the
raw message from the evaluator, let's add a `EvaluationError` structure
that implements the error interface so it can be tested for with `As`,
unwraps into base which is `ErrEvaluationFailed` through the usual
`NewErrEvaluationFailed`.

The user-visible result is that the errors are no longer stored with the
`evaluation failure` prefix, but raw, which will be useful for storing
JSON in the result.