Add a gauge to count the quickstart profiles (#1775)
The quickstart profile are a bit of a special case, with a special name
(that's not enforced, but I don't think the number of false positives
would be very high). We don't have another way to see if a profile was
created by the quickstart tool and while we could add one, it's not
worth the effort.

This PR adds a gauge that counts the profiles with the quickstart name
so that we can get metrics on their usage.