Remove commit lint job
While it seemed like a handy idea to lint commit, it turns out that the
job itself is not actually used by this project by the fact that we
squash the commits before merging. This bypasses the lint job in the end.

On the other hand, some of the guidelines like title lengths and commit
lengths ended up being controversial. We'll revisit this with a CI
job that checks for PR messages instead.