Always return response from github_rest's Do()
The Do method in our github_rest wrapper behaved a bit differently from
the go-github method it was wrapping - instead of always returning the
result so that the caller might check it, it would always return nil on
error.

This made checking for specific error codes (like 404) awkward
especially if we pass this interface to a library as you'd have to cast
the error to go-github's error codes and then grab the reply from there
anyway.

Passing the body always would enable us to use this interface (or a
variant of it) in frizbee instead of directly passing in the go-github
instance.