Don't `ExitNicelyOnError` on the sub-commands (#1889)
Explicit is better than implicit.

We already handled the output rendering in the cobra CLI root, which does
the explicit `Execute` function. This was already even done with the
`ExitNicelyOnError` construct. So, instead of handling the logic in several
places, this moves returns the error everywhere.

The caveat is that we would miss the caller's context. That is, we would miss
a message of what was going on when the error happened. To address this,
I introduced a function called `MessageAndError` which allows us
to print a message to the right stream in cobra and then return the error.