Add default branch as a repository field
This will be handy to stop hard-coding the branch in the Git-related
evaluations, we could just rely on the default branch for the project.

In order to facilitate adding/removing fields from the repo protobuf, this
also introduces a helper function that centralizes setting the values
from a database entry.