Fix pull_request remediation from branch (#1976)
Turns out that the `Head` parameter of the `PullRequestListOptions`
structure doesn't just take the branch name but `owner:branch`.

The tricky thing is that if you just use the branch name, then the list
operation would just return all the currently open PRs. This would
manifest as minder pushing some code into a branch, but then not opening
a PR if any other PR was opened against an enrolled repository.

I didn't notice this in my initial testing beacuse I would either have 0
PRs or just 1 with the tag-to-sha patch.