Add `type` to diff ingester protobuf (#1993)
The diff ingester was initially just coded to support a diff in dependencies.

Howver, this ingester is very handy for other things such as full diffs or
diffs of github actions.

This introduces a `type` to the protobuf which allows us to set the diff
type as needed.

This also introduces validation which makes sure that a user can only set
an empty or `dep` diff type. Empty keeps backwards compatibility.