Don't print informational message on stdout
The message that the config file is not present and that default are
about to be used should not be printed on stdout as it breaks json
output - you can't then just pipe the rest of the command to jq.

I would honestly prefer to remove the message altogether as asking users
to redirect stderr to /dev/null seems like a nuisance, but let's start
with not printing the message to stdout.