Make webhook tests more resilient
The webhook tests start an http server and spawn it in the background. This leads
to sporadic test failures, as the caller code might sometimes run before the
server is up.

This adds a retry with backoff in those tests, so if this happens, the caller will
simply retry.

If the retry fails, this will actually be a real issue with the code.