Don't execute actions when running `mindev ruletype test` (#2023)
* Don't execute actions when running `mindev ruletype test`

This overwrites the remediation and alert to be `off` since we don't want to
execute actions when testing. This is not _yet_ supported and causes a segfault.

So, let's turn it off for now. If the need comes to test the actions themselves
with `mindev` we can work on turning it on later. But, for now, no segfault
is better.

Closes: https://github.com/stacklok/minder/issues/1858

* don't disable remediations