Move EntityInfoWrapper to a separate package to avoid import loop
With the addition of the business telemetry, the engine will depend on
the telemetry package, but since we need to also manage entity info
wrappers in the telemetry, we would have a dependency fro the telemetry
package on the engine as well.

To avoid this import loop, let's move the entity info wrapper into a
separate package.