Don't default to `minder` as default store ID is authz configuration (#2163)
We should default to an empty string since we initially generate and
then list the store ID's when building the authz client. Passing a
default store ID is not the right thing to do. We should instead just
pass a store ID from the configuration once we have already bootstrapped
the app.