Actually run authz tests (#2165)
The authz tests we had were actually not issuing an error because the
`fga` command outputted the "Usage" and failed silently. It turned out
not to be trivial to run the command from our tests, so I ended up
adding the `fga` CLI installation to the bootstrap target, adding a
Makefile target and creating a CI job out of it.

The `authz` directory was moved into `internal/authz/model` in as a
cleanup. But also, we'll eventually embed the model in the go server as
part of the migration.