Add authorization JSON model workflow (#2166)
The OpenFGA API does not know how to handle the DSL natively and relies
on users tranforming the DSL into JSON. So, this adds a Makefile target
to do so. We also add the generated JSON structure into our repo.

The generation now runs as part of the general `make gen` target.

This also add a validation step that runs before the authz tests.