Add permissions API to FGA model (#2219)
This adds the permissions API to the FGA model. This API is highly
privileged, so only admins can take it into use. Note that we also added
a permissions manager role that's meant for SaaS-like cases where a
support engineer can provision a project for a user by giving them
access.

Note that uses of these future API endpoints should be heavily audited.