Add ability to dynamically allocate role assignments/grants
This adds the concept of claims mappings, which can be used to
pre-allocate role assignments when you want to add permissions for a
user that is not yet registered in minder. It adds a set of claims that
need to match from the token for the role to be resolved.