Move projects to be self-contained trees
This moves the projects users are enrolled to be self-contained trees.
This thus removes the need for organizations as a whole. We'll slowly
start moving towards removing this concept from the code-base.

For now, this removes the organization-related queries and makes users
point to their project instead.

User deletion is now handled by figuring out the projects a user had
grants in and removing the projects if there's no more tuples in them.

Actual removal of foregin key constraints for organizations will happen
in a subsequent PR.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>