Fix user deletion logic (first minder then keycloak)
This changes the user deletion logic to first delete a user and the
relevant project information in minder and only after that doing the
deletion in keycloak. This would address issues where the deletion gets
interrupted thus leaving a user's project in minder whereas keycloak
would provision an entirely new user.