Use latest golangci-lint in CI (#2365)
Instead of having to manually maintain the golangci-lint version in our
job, let's instead default to the latest one. This will get us the
latest and greatest CI checks provided by the tool and reduce the
maintenance burden of manual upgrades to this file. Unfortunately, this
is something that dependabot does not support updating.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>