Use only the status comment, no replies in review summary
This commit builds atop the previous one and moves further so that the
message in the review reply (not to be confused with inline replies) are
empty and all the feedback is kept in the single summary comment.

Because we still need to dismiss the reviews, but now we don't have a
review text body where we used to keep the magic comment with metadata,
we keep it in the summary comment instead.

We also keep the review ID there because we can't be sure which is the
minder review now in case the same identity is used for a human identity
and minder.