move profile validation logic out of `controlplane` (#2402)
* move profile validation logic out of `controlplane`

Currently there is a lot of logic in the grpc controlplane which is not
always easy to test. As a refactoring exercise, I lifted profile
validation out of the controlplane and placed it inside its own struct.
I also took the opportunity to add some extra test coverage to this
logic. Part of my motivation is to create some examples and discussion
about how we might structure the codebase as we add more business logic
(or refactor existing logic).

* cleanup error handling

* fix lint issue