Move webhook create/delete into separate interface
Part of ongoing test improvement work

This started out as an attempt to refactor
`registerWebhookForRepository` (which currently is the least-tested
method in the codebase by # of untested lines). I discovered that
breaking it up will require further rounds of refactoring, but I started
by moving webhook logic behind a separate interface, since it was
probably the easiest part of the logic to isolate.

Existing tests for `handlers_repository` were altered to use the new
interface which simplified those tests somewhat.