promote PR ID and line number to int64
Fixes #2443

PR ID and line number are were previously downcasted to int32 in a
number of places despite being mostly treated as an int64 type. While
the likelihood of either of these values exceeding a 32-bit MAX_INT are
minimal, this change makes them 64-bit to err on the side of caution.