Move repo creation/deletion into separate interface
This relates to ongoing test/refactoring work.

Move the methods for repo creation and deletion behind their own
interfaces, and add test for the new code and the controlplane endpoints
which use those methods.

In future, we may want to incorporate this Repository interface into the
providers somehow so that the controlplane can create repos for
different types of source control systems without needing to know which
exact type they are.