Upsert provider access tokens instead of Create and Delete
For OAuth flow provider enrollments we used to delete and recreate the
provider access token in order to be able to list the new token based on
CreatedAt. For the case where we pass a PAT, we would just error out
in case the token was already created.

Let's just upsert in both cases and instead of CreatedAt let's look at
UpdatedAt.

This will be useful for tests.

Fixes: #2411