Move repository create to RepositoryService (#2632)
This completes the current refactoring of repository logic.

One small behaviour change is introduced: when the create logic is
unable to insert the new repo into the DB, it will attempts to delete
the webhook which it created. The previous logic would just leave the
webhook in the customer's github repo.