Sort the list of profiles before returning to clients
Because we internally construct a go map when retrieving the list of
profiles and then iterate over it, the result of ListProfiles is not
ordered. But this is putting a strain on the UI that would have to order
the results or else every refresh might produce a different looking set
of results.

Fixes: #2658