Move profile create/update to a separate interface (#2653)
Fixes #2541

Move the create and update methods into a separate package and
interface. Unlike my previous PRs which split out code into separate
interfaces, I have decided not to write new tests for the new
ProfileService. This is due to the number of test cases which are needed
for profiles. Instead I am relying on the existing profile handler
tests in the controlplane which at least cover the creation workflow.

Changes to the logic in the create and update methods have been kept to
the bare minimum needed to split it out into its new home.

Changes verified locally using the CLI.