Add label support for ListProfilesByProjectID
In order to support listing by labels, let's extend the
`ListProfilesByProjectID` SQL statement with two optional parameters -
`include_labels` and `exclude_labels`. These default to `[]` which is
also the first branch in the SQL statement which would list all profiles
with no labels.

Please refer to the unit tests for more use-cases.

Related: #2759