Explicitly coalesce labels on create and update to {} (#2794)
I was stupid and thought that setting the default on the column in
migration would ensure that we default to '{}' as well as the
grpc-generated getter returning an empty slice by default. Turns out
that one defaults to returning nil.

To fix that, we explicitly rewrite nil labels to an empty array.