Modify GRPCClientWrapRunE to pass positional arguments (#2967)
Relates to #2966

The current implementation swallows positional args. Change to pass them
to the command functions.

I considered creating a second function with a different name which would
pass arguments, and thus leave the current code alone. However, I felt
this would cause confusion long term, and that changing the existing
code to the new interface was trivial work.