Add a unit test for switch of a profile state to pending
We recently suspected that one of our stored procedures is not behaving
as it should. This turned out to be false, but during debugging of the
stored procedure I already wrote a unit test for the behaviour that was
not working at the moment.

It makes sense to submit it in case we regress in the future.