Remove hardcoded dependency between artifacts and projects
This makes repos nullable in artifacts. As of today, this is not really
a case that actually happens... so this should effectively be a no-op.
However, the idea is that soon we'll start getting artifacts that that
come from other providers that are not the same as the repository they
are linked to.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>