Remove dependency on provider name in repository table
Now that repositories has a foreign key reference to the providers
table, it is not necessary to store the repo name (and indeed, it
introduces the possibility that the names could go out of sync).

In order to get rid of the provider name column, I have decided to take the
following approach:

1) Created a local migration script to drop the column
2) Changed the queries, code, tests accordingly.
3) Got rid of the migration script.

The idea is to pretend that we have deleted the column, but not actually
delete it. This will make it _a lot_ easier to rollback this change if
something goes wrong. If all is well, we can delete the provider name
columnn in a future PR.

Quite a few places in the code expect to be able to get the provider
name from a repo row. In some cases, it was straightforward to change
those pieces of code to use the provider ID. Removing the provider name
coupling from some parts of the code (particularly relating to event handling)
is a much bigger job which needs to be done in a separate PR. To avoid
changing those parts of the code, I have changed some of the
repository-related queries to return a tuple of provider name and
repository - this allows those call sites to continue to get a provider
name without introducing additional queries or making invasive changes.