Issue a bespoke profile update message on patching a profile (#3012)
I forgot to issue a profile update message on patching a profile. Since
the profile update message is currently nicely encapsulated in
`sendNewProfileEvent` which is a private method of the Profile service
and we already have PR #2990 that would move all of the Patch
functionality to that service and remove the current Patch handler
implementation, I decided to create a bespoke message for now.

This code will be removed once we merge PR #2990.