Use ProviderID in telemetry data (#3026)
Relating to my PR for repositories, and Ozz's PRs for artifacts, add
ProviderID to the telemetry struct. This is because there will be
certain situations where we will not know the name, but know the ID
instead.

I have changed some of the code to use ProviderID instead of Provider
name where it made sense to do so. In some places, it is not possible
yet, so I have left comments. In the controlplane, where the provider
name is an input via the request context, I have used both - tracking
the provider name input may still be useful for debugging.