Fix CLI doc generation bug
Fixes #2626

Actual cause of this issue is that MDX A) needs some of the characters
in the sample shell code to be escaped and B) even if you escape it, it
wont render the line as markdown properly.

Do some unholy things with sed to make it work. Forgive me.