Use provider ID instead of name when sending events
Relates to #3071

This PR changes any code which creates an EntityInfoWrapper to use
Provider ID instead of name. The name is still left as a field in the
event for backwards compatibility purposes, and the consumer can still
use the name to create the provider instance. However, the name is no
longer populated.

Some other refactoring was carried out, specifically moving code from
the helper package to more appropriate packages.