Ignore description and defaults when validing ruletype updates
We've been doing a bunch of changes to the defaults and descriptions to
make UI render better, but those are being flagged by minder ruletype
apply as incompatible because we require that the new ruletype is a
strict superset of the old one.

Let's ignore the description field as long as it's a scalar with a string
value and let's ignore the default field as long as it has a 'type'
sibling. These additional checks make sure that we don't accidentally
allow updating parameters that are named default or description.

Fixes: https://github.com/stacklok/minder/issues/3079