Add ProviderID field to EntityInfoWrapper (#3072)
Relates to #3071

This PR adds the ProviderID field to the event structure, and changes
the executor to use the ProviderID if present. Note that the ProviderID
is not actually set anywhere at this point: I am doing this changeover
in phases to avoid backwards compatibility issues as we change the event
structure.