Add `minder auth token` command (#3107)
* Add `minder auth token` command

This command allows you to retrieve a token from the identity provider
used by the given configuration. This is handy when trying to do work on
a direct API, for debugging, or for working with other tools in the
ecosystem (e.g. trusty).

It'll try to authenticate and print out a token to be used, or simply
read the token that's already in `credentials.json`.

An ideal flow would be `minder auth login` then `minder auth token`.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>

* fix copyright year

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>

---------

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>