Return 500 if Github AppName is empty when enrolling provider (#3124)
I did not define this in my local configuration, and it led to some
weird errors when enrolling a Github App provider - specificially, it
creates a link with an empty app name which will lead to a 404 when the
link is acceessed. This will at least return a 500 if the value is omitted,
which makes the proble more explicit.