Don't use SELECT * when joining profile tables (#3130)
This patch has no functional changes, it is just a cleanup in our Go
database models that I didn't want to send before the recent release.

If sqlc generates a model for a query that joins over two tables, it
autogenerates column names that are represented as structure members, so
you end up with a single structure with members such as `ID` and `ID_2`.

This can be confusing and invite bugs. It's much cleaner to use
`sqlc.embed` to make sure that each table is represented by a separate
attribute in the models.