Make provider class mandatory in DB (#3132)
Relates to #2845

Undo migration #35 after discussion with Ozz and Jakub this morning. We
want the provider class to store the concrete type of the provider, this
will be used by the ProviderFactory.

After studying the code, there does not appear to be any situation in
which this field will be null in the staging/prod DB.