Add external IDs to entities
Relates to #3142

Per our design discussions, add an opaque identifier column to each
entity type. This is intended to contain provider-specific information
used to identify that entity in the provider's API. This will allow us
to build generic interfaces around providers which do not need to know
exactly what the provider needs as an ID for each entity type.

Alternatives considered:

1) Use a string. The problem with this approach is that since the string
   can be inspected, it creates the temptation to write code outside of
   the provider which uses the structure/contents of that string. This
   is meant to be an opaque identifier.

2) Use a UUID, which serves as a foreign key reference to data in the
   provider's database. I considered this for a while, but it creates a
   requirement for each provider to have a database. This may be a
   reasonable assumption for many cases, but I think we should leave the
   door open for stateless providers when carrying out simple
   integrations.