Use `go generate` directives for mock generation
Instead of adding commands to the Makefile, use go generate directives
on top of the interfaces we need mocks for. This approach has the following
advantages:

1) Less context switching: you can declare the need for a mock in the
   code and not have to edit the makefiles.
2) The same directive can be reused everywhere - you no longer need to
   specify the filepath, interface name and the path to the generated
   mock.
3) If the interface is renamed or moved, the code generation will still
   work without modification.

I have left some commands in the makefile for generating mocks from code
in `pkg`, and from autogenerated code. We could in theory get rid of
these usages as well, but I did not feel there was much point at this
time since they are edge cases.