More descriptive error message when validating secrets with any of the fallback webhooks fails
Previously we would have returned a nil error but also a nil payload on
exhausting all previous secrets in a loop which would later error out when it
couldn't be parsed. That means that this bug has no security implications (we
would just get a 500 instead of a 400), but let's fix it nonetheless.