Refactor repo deletion to move db/provider logic behind interface
Relates to #2845

Change the interface of the repo service so that it takes care of DB
queries and provider instantiation. This moves various details about
repo cleanup out of the controlplane.

As part of this, change the provider cleanup process to delete webhooks
directly instead of deleting repos. This avoided a circular dependency,
and more closely aligns with the idea that webhook management is a
provider-specific detail.