Implement GitHubClientFactory (#3203)
Relates to #2845

For a future PR, I want to be able to easily instantiate the GitHub
client without going through the provider creation process. This splits
the code for creating the github client out into its own factory. As a
bonus - this removes some code which is duplicated between the oauth
and app provider instantiation flows.