Ensure provider name is set during repo delete
This was an inadvertant change introduced by my previous PR which
refactored the repo deletion flow. Prior to my PR, the code implicitly
validated that the provider name was non empty by setting `Valid` to
true in the `sql.NullString` used to query repos by provider name. This
change restores that behaviour, adds comments, and also adds explicit
validation in the controlplane to ensure that provider name is set.

Since an empty provider name would always have led to an error, I do not
believe that this extra validation constitutes a breaking change. It
will however change the response code in this case from a 404 to a 400.