Pass RestClientCache as constructor param
Relates to #2845

This has been broken out of a larger PR for ease of review.

The RestCacheClient is used in several parts of the code, but is always
passed as an option. In some places (which will be changed in a
subsequent PR) this is used as a form of dependency injection. However,
this pattern has caused problems in some of the refactoring I am doing.

This PR changes the code so that it is passed around explicitly to the
constructor methods of the structs/interfaces which need it. This is an
interim step - the desired end state is that the RestClientCache is
hidden from most of its current clients, and only the GitHub
ProviderClassManager will depend on it (along with the Github trait
instances it produces).