Refactor application wireup
Relates to #2845

With some of the changes I've made, I have started to run into issues
with the way we currently wire up the server, event handler and
reconcilers:

1. The way in which structs are shared across events, server and
   reconcilers is a bit hacky, and will get more hacky when I start
   sharing more structs.
2. The functional options pattern is starting to cause complexity now that we
   need to wire up dependencies which depend on some of the structs
   passed by the options.

This moves most of the application wireup outside of `server.go` and
into `service.go`. The optional parameters for the server are removed
entirely, and are now passed in as explicit arguments. Various tests
have been cleaned up to instantiate the Server directly with appropriate
stubs for different dependencies.

The use of optional parameters for the reconciler and event handler will
be cleaned up in a future PR.