Make Project Create/Delete a service (#3227)
Relates to #2845

This is part of a change which allows me to avoid some ugly hackery in
my draft PR (#3221). Break Project creation/deletion out into a pair of
dedicated structs/interfaces. Change the `ProjectFactory` function used
in the GitHubProviderService from a method on the controlplane to a
function which returns a function. This is needed for some additional
cleanup introduced in a future PR.